requirements.txt
aiohttp~=3.8.4
aiofiles~=23.1.0
pydub~=0.25.1
sqlalchemy[asyncio]~=2.0.4
asyncpg~=0.27.0
psycopg2~=2.9.5
aiosqlite~=0.17.0
alembic~=1.10.1
aiogram~=2.25.1
JSON-log-formatter~=0.5.1
redis~=4.5.1
# nest_asyncio~=1.5.6
aiocloudpayments~=0.1.1
voximplant-apiclient~=1.3.7
boto3~=1.26.118
python-dateutil~=2.8.2
asyncio~=3.4.3
pymorphy2~=0.9.1
botocore~=1.29.118
phonenumbers~=8.13.35
